Begin implementing the exchange CLI
This commit lays out the cli of the exchange program

It creates the file from the publisher csv and does proper error checking
It also loads the generated exchange csv into a merkle tree.

Since the csv we generate for the exchange has less fields than the
one for the publisher we had to make a new struct and new functions.
We should remove this duplication and just remove the extra feilds
from the publisher csv