Start a super basic tcp listener
This will allow us to test and send commands to the publisher and verifier.
The shell and the listener do not interact directly with each other,
rather, the common functions will be moved to a separate file, which both
the shell and listener can use distinctly.
The listener will be responsible for its own argument parsing logic,
much of which will most likely come from the shell anyway.