Refactor the publisher and exchange shell error handling
The main drivers behind this are a new ArgsList struct and sanatize_args
macro. These allow us to more easily hdndle no-arg and empty/invalid
argument cases.
As part of this, this commit allows the shells to continue with an
error if they are unable to connect to Redis.
There are still a lot of areas within the code where we could do better
error-handling, but this gives us a start, and most likely, we can
apply similar ideas from here