Handle a race condition in exchange db creation
When writing the exchange file, if the file is created and deleted in a
loop as part of a cargo test test=case it may race with the
test that only creates the file and then deletes it. This is because
cargo test runs tests in parallel.
To address this we acquire a lock when creating the csv file and drop it
when the function is finished.
It is unclear if this scenario could happen with normal usage, but this is a
small precaution