Make merkle tree optional for the exchange shell
Mirroring the same behavior as the publisher, the exchange makes a
merkle tree from the provided csv file path.
However, that responsibility for the exchange will now be handled by emp
and the generated c++ script file(s). Since that part is not fully done,
we mark the merkle tree as optional to be able to test the new txt
loading path that should be merged with our csv logic